重构检查表

+++ src/main/java/com/ts/server/safe/base/service/UniCheckTypeService.java
@@ -2,8 +2,8 @@ package com.ts.server.safe.base.service;
 
 import com.ts.server.safe.BaseException;
 import com.ts.server.safe.base.dao.UniCheckItemDao;
-import com.ts.server.safe.base.dao.UniCheckTableDao;
-import com.ts.server.safe.base.domain.UniCheckTable;
+import com.ts.server.safe.base.dao.UniCheckTypeDao;
+import com.ts.server.safe.base.domain.UniCheckType;
 import com.ts.server.safe.common.id.IdGenerators;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
@@ -14,62 +14,60 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 /**
- * 统一检查表业务服务
+ * 检查类别业务服务
  *
  * @author <a href="mailto:hhywangwei@gmail.com">WangWei</a>
  */
 @Service
-@Transactional(propagation = Propagation.REQUIRED)
-public class UniCheckTableService {
-    private final UniCheckTableDao dao;
+@Transactional(readOnly = true)
+public class UniCheckTypeService {
+    private final UniCheckTypeDao dao;
     private final UniCheckItemDao itemDao;
 
     @Autowired
-    public UniCheckTableService(UniCheckTableDao dao, UniCheckItemDao itemDao) {
+    public UniCheckTypeService(UniCheckTypeDao dao, UniCheckItemDao itemDao) {
         this.dao = dao;
         this.itemDao = itemDao;
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public UniCheckTable save(UniCheckTable t){
-        if(!itemDao.has(t.getId())){
-            throw new BaseException("关联检查项目不存在");
-        }
+    public UniCheckType save(UniCheckType t){
         t.setId(IdGenerators.uuid());
+
         dao.insert(t);
+
         return dao.findOne(t.getId());
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public UniCheckTable update(UniCheckTable t){
-        if(dao.update(t)){
-            throw new BaseException("修改检查表失败");
+    public UniCheckType update(UniCheckType t){
+        if(!dao.update(t)){
+            throw new BaseException("修改检查类别失败");
         }
         return dao.findOne(t.getId());
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
     public boolean delete(String id){
+        if(itemDao.hasType(id)){
+            throw new BaseException("有所属检查项目不能删除");
+        }
         return dao.delete(id);
     }
 
-    public UniCheckTable get(String id){
+    public UniCheckType get(String id){
         try{
             return dao.findOne(id);
         }catch (DataAccessException e){
-            throw new BaseException("检查表不存在");
+            throw new BaseException("检查类别不存在");
         }
     }
 
-    public Long count(String typeName, String itemName, String content, String lawItem){
-        return dao.count(typeName, itemName, content, lawItem);
+    public Long count(String name){
+        return dao.count(name);
     }
 
-    public List<UniCheckTable> query(String typeName, String itemName, String content, String lawItem, int offset, int limit){
-        return dao.find(typeName, itemName, content, lawItem, offset, limit);
-    }
-
-    public List<UniCheckTable> queryOfItem(String itemId) {
-        return dao.findOfItem(itemId);
+    public List<UniCheckType> query(String name, int offset, int limit){
+        return dao.find(name, offset, limit);
     }
 }
